requirements.txt
Django==3.0.8
PyJWT==1.7.1
amqp==2.6.0
anyjson==0.3.3
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.6
coverage==5.2
ecdsa==0.15
flake8==3.8.3
gunicorn==20.0.4
kombu==4.6.11
logilab-common==1.7.2
mccabe==0.6.1
meld3==2.0.1
mock==4.0.2
paramiko==2.7.1
pep8==1.7.1
pip-tools==5.2.1
pycrypto==2.6.1
pyflakes==2.2.0
pylint==2.5.3
pytz==2020.1
redis==3.5.3
setproctitle==1.1.10
six==1.15.0
supervisor==4.2.0
